+++ CHANGES.rst
@@ -28,6 +28,7 @@
 * Added Python 3.6 compatibility
 * Removed the unused ``nullstr`` and renamed ``nullunicode`` to ``nullstr``
 * New: ``add_primary_relationship`` to define a primary child on parent models
+* Added ``NoIdMixin`` that is BaseMixin minus the id column
   
 
 0.5.2